add user center.

+++ src/main/java/com/zuoxiaolong/servlet/UploadImage.java
@@ -1,21 +1,19 @@
 package com.zuoxiaolong.servlet;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.List;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-
+import com.zuoxiaolong.config.Configuration;
+import com.zuoxiaolong.util.IOUtil;
+import com.zuoxiaolong.util.ImageUtil;
 import org.apache.commons.fileupload.FileItem;
 import org.apache.commons.fileupload.FileUploadException;
 import org.apache.commons.fileupload.disk.DiskFileItemFactory;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.apache.commons.lang.StringUtils;
 
-import com.zuoxiaolong.config.Configuration;
-import com.zuoxiaolong.util.IOUtil;
-import com.zuoxiaolong.util.ImageUtil;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import java.io.File;
+import java.io.IOException;
+import java.util.List;
 
 /*
  * Copyright 2002-2015 the original author or authors.
@@ -50,6 +48,11 @@ public class UploadImage extends AbstractServlet {
 			List<FileItem> items = upload.parseRequest(request);
 			FileItem fileItem = null;
 			if (items != null && items.size() > 0 && StringUtils.isNotBlank((fileItem = items.get(0)).getName())) {
+				String fileName = fileItem.getName();
+				if (!fileName.endsWith(".jpg") && !fileName.endsWith(".gif") && !fileName.endsWith(".png")) {
+					writeText("format_error");
+					return;
+				}
 				path = ImageUtil.generatePath(fileItem.getName());
 				IOUtil.copy(fileItem.getInputStream(), Configuration.getContextPath(path));
 				fileItem.delete();
